Closing the overlay reset the position to center of window
Initial position defined using style get applied only on first display. Closing the overlay reset the position to center of window. This was happening due to resetTargetDimensions removing the styles on overlay target.
Added code to capture the initial dimensions of target on domReady and targetChanged and reset to saved dimensions.